Add protoc-native way of option handling
The protoc tool has the generic mechanisms for plugin option handling.
The plugin-specific options can be specified using the
--<plugin_name>_opt argument or by prepending options to the
--<plugin_name>_out argument with colon separator. In both cases
options need to be separated by semicolon. Add the processing of this
argument to generator and use --<plugin_name>_opt as the default one
in CMake API. Also adjust format of the QT_PROTOBUF_OPTIONS environment
variable to make them consistent.

Change-Id: I269ca7b9b2a952d0d30b18cb57c1465ae261c83a
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>