Remove QAbstractGrpcService and disable QtGrpc service generation
It's not in use for now, so there's no reason for it to exist in
the public API, even when we're in TP - and plans for future
releases are still unclear regarding server-side gRPC support,
so this feature may not be needed in the coming months.

Either way, it can easily be added again once we have all the
server functionality ready.

Split code generation of gRPC client and server to separate
functions and make GenerateServerService unused.

Task-number: QTBUG-108833
Change-Id: I65a2122ec5dba640614d97daec973567c57401f4
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 3d16e50a4bc9e91fe7e9b8716a30c9b919861321)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>