Extend qt_internal_add_protobuf_module function
The function now accepts the list of .proto files as an argument and
calls 'qt6_add_protobuf' implictly. .proto files are treated as an
interface of the module and installed to the module include path.
The respective module include path is added to the target's
QT_PROTO_INCLUDES property for the further use in user projects as
PROTO_INCLUDES argument of 'qt_add_protobuf' and 'qt_add_grpc'
functions.

QtProtobufBuildInternals.cmake is renamed and moved to the correct
directory for the implicit processing by qt_internal_add_module
function.

Pick-to: 6.5
Change-Id: I3322a18a6c487bb8dd0165ce9bca5f84c89cbbd5
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>