Introduce QProtobufMessagePointer and un-virtualize QProtobufMessage
Changing our return types from QProtobufMessage* to use a smart-ptr lets
us quite nicely get rid of the only virtual left in QProtobufMessage:
The destructor. Since we create the subclass through the meta-type
system we can also delete it that way. This gets rid of the virtual
table, which we would've been unable to extend anyway given our ABI
promises.

Since users can no longer call `delete` on a QProtobufMessage directly
we instead provide a convenient alias for a unique_ptr with a Deleter
which takes care of the aforementioned meta-type part.

Done-with: Marc Mutz <marc.mutz@qt.io>
Pick-to: 6.5
Change-Id: I3405f0a9121dc8325f2940a8ee719c3c462f2c01
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>