Fix the way proto includes assembled into a single command
The CMake list(JOIN command doesn't expand the generator expressions
before merging the list. This leads to the situation when generator
expressions that are empty string after the evaluation put an empty
-I argument to a protoc command line. Using '$<JOIN ' generator
expression we avoid this situation since list will be assembled into
argument only after evauation of the all nested genertor expressions.

Pick-to: 6.5
Change-Id: Ibbb141a46e89bb71c6f6151c70b7ec754b44355f
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>