Isolate host tool dependencies from the cross-compiling environment
The Qt6ProtobufTools and Qt6GrpcTools expose their dependencies to
a build tree. The created targets then are treated as the dependencies
for the target tools that cannot link to them. To avoid making these
dependencies visible for the whole project moving them to a
subdirectory that hides them from target tools. Also avoid using
cmake package registry when looking for these packages since cmake
knows nothing about the way we cheat when looking for the host tools,
and thinks that the packages were found previously are compatible with
target tools.

Fixes: QTBUG-111140
Task-number: 6.5
Change-Id: I25bc6c687af0be1f0549226eff556c22afc1b7f0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>