CMake: Fix commands not being available when cross-building examples
There were a few iterations of trying to ensure that protobuf and grpc
tools are available when cross-building Qt itself, but do not fail
configuration of a cross-built Qt if the host tools are not present.

The latest iteration ensured that worked fine when it built Qt itself,
but it did not account for still finding tools when cross-building
user projects (iOS / Android).

That was meant to be done by calling qt_internal_add_tool() ->
qt_internal_find_tool(), but the latter exits early if the tool is
already found, which is the case for qtgrpc because it explicitly
looks for tools in the src/tools/host sub-project, and thus
qt_internal_append_known_modules_with_tools is never called and a
dependency on the Tools package is not written into the iOS version
of the Qt6ProtobufDependencies.cmake file.

Until qt_internal_find_tool() learns to deal with that, we can
register a dependency on the Tools package manually by using
qt_record_extra_qt_main_tools_package_dependency() similar to how we
do it in qtwayland.
We only register the dependency if we the tools are not being built in
the current build (which implies a cross-build and that the
tools are already found).

Amends 0ae19cc745384cc8c6fc07e6bf2d3e163f3f61d7
Amends 39f2a1e64a52720bf758b5a7e5444a82656d6a18

Pick-to: 6.5 dev
Fixes: QTBUG-111983
Change-Id: I2c37b89fd960aa142b8c4a5f150ad2067501f78c
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>