Fix the deserialization of non-packed lists
This fixes the situation when we received the non-packed repeated field
from the wire, but the used protobuf schema specifies the field as
packed. This situation is possible if the protobuf schema mismatches on
on the wire ends. The explicit indication of this situation is the
wiretype mismatch. So if the wiretype differs of the expected one that
means we need to look for the handler with the 'packed' option opposite
to the one is expected by the actual protobuf schema.

Pick-to: 6.5
Change-Id: I57bf613410c4031f638533f092fd8d36c592db32
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>