Cleanup protobuf serializer
The actual protobuf serializer has multiple overheads and code
duplicaiton. Serialization functions use individual ways to
detect if field was initialized or not. This patch adds the
general functions that detect if field was initialized or not.

Also serializers previously juggled the protobuf field number to
conclude the need of field header. Now the header is added under
clear conditions.

For STRING and BYTES lists we should use the same way of
serializing/deserialing as for non-packed lists, since they are
non-packed lists.

Map elements should be serialized even if value is not initialized.
But non-initialized keys should still be present in serialization
stream. Avoid using serializeProperty when serializing the map key
and use integated type serilizers directly.

This change is backported from the dev branch, but is significantly
different of the original one, since the code base of dev branch
contains several features, like the support of the oneof fields,
that were not backported to the 6.5 branch.

Change-Id: I32b512acdaa9f355eea9c41f735bc9722bed0bba
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 39e777316028f3f41078801c6cea6fc2dc9add84)