Reuse the existing message value when deserializing it second time from the wire
As per the conformance "Merge" test expectations we should reuse the
existing message value and merge values received from the wire or
override existing. Check if the value stored in the property of the
message type is not 'null' and reuse the message stored when
deserialing the property more than once.

[ChangeLog][Important Behavior Changes] Initialized message fields of
the message type are reused when deserializinng the message. Previously
message fields of the message type were deserialized to a freshly
created message object. This change removes creating of the new message
object at each deserialization cycle and reuses the exising field value
if it's not 'null'. If fields of this message object are received from
wire they will be overritten, if not they will store the previous value.

Fixes: QTBUG-112421
Fixes: QTBUG-112422
Fixes: QTBUG-112429
Change-Id: I73a792e72064af68b5a2a1e61e944c52a79eb412
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>