Doc: Update configuration for Qt GRPC/Protobuf examples
Due to infrastructure issues (specifically, missing replication of
repositories to code.qt.io), Qt GRPC and Qt Protobuf relied on falling
back to QDoc's built-in generation of lists of files that comprise an
example rather than a link to code.qt.io, the latter of which is the
preferred option for examples in Qt. Said list limits its output to
files identified by the `examples.fileextensions` QDoc configuration
variable. However, the `.proto` file extension was not added to the
documentation configuration for these modules.

Since the module released, repository replication issues have been
resolved. Thus, it is now possible to link to code.qt.io, as preferred.

This change adds the `.proto` file extension to the configuration,
and sets the `url.examples` configuration variable such that QDoc will
generate a link to code.qt.io. The file extension change is still
considered relevant, in case the fallback is required in the future,
and `.proto` is not part of the default set of file extensions for Qt's
examples.

Fixes: QTBUG-112754
Pick-to: 6.5
Change-Id: I61fcc1fac0aa11a4dcf278deabbd7c12c091d894
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>