QtProtobufQtTypesQtCoreTest::qDateTime(LocalTime) fails
LocalTime test-case is refactored.
Idea of particular test is serialize/de-serialize object and verify
with expected constant result.
But in case LocalTime is used for QDateTime creation, the resulting
QDateTime object will always be different depending on system time zone
of concrete user.
So it makes sense to preset timezone before LocalTime object creation,
and returned it back after.

Fixes: QTBUG-113342
Change-Id: Ic4dfd033ff00d525225669fe771de6fb386d1b0b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>