Make sure we are able to compile and link simple app using libprotoc
There are no guarantees that the protoc library found in the system
is compatible with the flags/libraries that we use when building
Qt tools. We may avoid compile-time error by checking this at configure
time. Add the configure time tests that check if the simple apps can
be compiled and linked to libprotoc and libprotobuf.

Fixes: QTBUG-115032
Pick-to: 6.5 6.6
Change-Id: Iaff157408bc131d472946cf0442ca634efab13b0
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>