Rework the Qml plugin export macro
Since the plugin is not exposed by any header file it doesn't make
sense to overcomplicate its export. Use internal cpp macro to export
required plugin. Also this fixes potential issue when export macro
is not provided by user to qtprotobufgen, so the export is not
generated and plugin symbols are simply unaccessible.

Pick-to: 6.6
Change-Id: Ie54cd94eb5ac9e6bb895c02456a6d6d6669a1daf
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>