Fix detecting of unused nested messages
Previously if nested message was never used by the nesting class it
was ignored to be generated. It was the result if invalid filtering
of map messages as nested types.
Use the explicit map_key() based check instead of weird and tricky
one based on nesting class fields.

Keep the fallback solution that uses MessageOptions when building
with libprotoc versions less than 3.12.

Fixes: QTBUG-115175
Pick-to: 6.5 6.6
Change-Id: I4aa238dac3375bdcdf85e41ba9daef74474ed0fb
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>