Change the way how generator detects enum scope
Enum visibility depends on containing type only. Other checks
doesn't make sense. If containing type is nullptr, this means
enum is not nested and is located in global scope, so it's GLOBAL_ENUM.
If containing type is the same as the scope, the enum is located inside
the scope, so it's LOCAL_ENUM. Otherwise enum is located in nearby
message and should be treated as NEIGHBOR_ENUM.

Pick-to: 6.5 6.6
Change-Id: I02e27f03f9eb5f28947d78c15855b83c7353efa0
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>