Overload getScopeNamespace
getScopeNamespace should use the message tree but not pre-defined
namespace string to provide the valid scope namespace. Overload
the existing function. The new function iterates over the type
containing types to find the scope. If scope is not found among
containing types then full nesting scope is used. If type and
scope are in different packages or scope is not defined(nullptr)
then full namespace is returned. New function also takes into
account that messages locate in namespace but not inside the
containing type.

Pick-to: 6.5 6.6
Change-Id: Ia807bb16676ca7791022c3e949b7f884c4131fee
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>