Filter the brackets from .proto file content
The recent finding from CMake shown that if list contain the
open bracket '[', CMake ignores any semicolon separators as this
list chunk is treated as escaped string until the closing bracket
is faced. This happened with .proto files that have the open
bracket at one line and the closing bracket on another one, or
even miss the closing bracket.

Remove all brackets preliminary from .proto file content to make sure
we don't hit this issue.

Note: We are not interested in brackets in this parser.

Pick-to: 6.5 6.6
Fixes: QTBUG-116043
Change-Id: I282e1f4a93db4486fde0370045c22d1c8ecf9afa
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>