Do not capitalize the enum field names
This lead to name conflicts that are not detected by protoc. Protoc
disallows using the same enum field names in single scope, but names
are case-sensetive. So we shouldn't capitalize the enum names since
it breaks the code in situations when two enums have values with same
name, but only case of first character differs:
message A {
    enum E1 {
       val = 1;
    }
    enum E2 {
       Val = 1;
    }
}

[ChangeLog][Protobuf][Generator] Enum fields now preserve the case of
first letter. Previosly generator capitalized the generated enum field
names.

Change-Id: I476c940041454918c958a4b31787731cadf9f6fe
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>