Remove use the use of weak_ptr from channelOperation
QMetaObject::Connection objects now control the channelOperation
lifetime through lambdas. We simply make sure that all connections are
destroyed when the communication is finished(by error or normally).

This is not obvious thing, but it allows to make sure that objects
won't be accessed after removal in lambdas, since they hold the
shared_ptr's now.

Pick-to: 6.6
Change-Id: Ie2715fe85c8ab9dc310f6212f85b0d02cfa5f561
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>