Avoid deleting oneof pointer if it points to oneof QVariant data
If oneof message field is already present, we return the pointer
to the QVariant internals in the QProtobufOneof::value method. In
this case deserializer accesses the message that is stored in the
oneof field directly. So if pointer received in the oneof field
setter matches the one that is stored in oneof field we should omit
the setter completely, especially the "delete" part since it will
attempt to free the QVariant stored data.

Amends e9c80d43f72a5e92e19536e574f0cef794ef0cb4

Pick-to: 6.6 6.7
Task-number: QTBUG-118996
Change-Id: I08a3fb0d52cdcd4ed2156b8e2c5e43244394005d
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>