Generate the explicit presence flag for protobuf message fields
Protobuf messages should have the explicit presence according to the
field presence specs. We should mark them with the respective field
flag to enable QProtobufMessage::property logic that verifies the
field presence according to the 'has<FieldName>' property value.

Add the ExplicitPresence field flag that now is generated for all
optional, message, and oneof fields.

[ChangeLog][Protobuf] All fields that have the message type now
generate the has<FieldName> property that indicates whether the field
is intialized or not.
[ChangeLog][Protobuf] All fields that have explcit presence now set
the 'ExplicitPresence' flag in their property ordering structures.

Task-number: QTBUG-120931
Change-Id: I962c38e1ebde04d393cafaa8b512a277d660b5fa
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>