Attempt to reconnect to server when calling any method
Attempt reconnecting to the server each time when calling any method.
Also clean the operations queue to avoid handling of the formally
'dead' operations when reconnection attemt was successful.

This fixes the general connection scenario that was missing. After the
first connection attempt channel never tried to connect again. This led
to regression introduced during the QGrpcHttp2Channel refactoring.
After making a call QGrpcReply stuck in noop state, since channel only
puts the assigned QGrpcChannelOperation to queue. This behaivor is
fixed by attempting to re-establish the connection to server if
QGrpcHttp2Channel is in the Error state. All the assigned
QGrpcChannelOperation will then either emit errorOccured signals if
connection is failed again or will complete the call if connection
is established.

Fixes: QTBUG-122677
Pick-to: 6.7
Change-Id: I92b8a1a707c5cb9e955d72100b19095564310a63
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>