Make the QHttp2Stream pointers raw pointers in QGrpcHttp2Channel
Store the QHttp2Stream pointers as the QPointer in QGrpcHttp2Channel.
The memory of the streams is managed by the QHttp2Connection
implicitly, so we should make sure that we do not attempt to remove
the stream that was removed by QHttp2Connection.

We cannot use raw pointers here because we pass Http2Handler to lambdas
and assume that Http2Handler lifetime is binded to the QHttp2Stream
lifetime.

Also ensure that we clear all Http2Handler before deleting the
QHttp2Connection instance.

The previous implementation led to double-free issue since,
QHttp2Connection was deleting all QHttp2Stream s that belong to it
first, and then there was semi-explicit removal invoked by
std::unique_ptr.

Pick-to: 6.7
Fixes: QTBUG-122908
Change-Id: I659eb9fc3f33dccc3ce70ba05db61f58973ce7c7
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>