Add the missing END_STREAM flag to the single-request operations
Single-request operations should send the END_STREAM flag to the server
in the empty DATA frame, that follows the last request DATA frame.
Add ChannelOperation structure to track the additional
QGrpcHttp2Channel related flags and options. The current implementation
add the endStream flag to the structure to ensure that unary calls and
server streams send END_STREAM flags in the empty DATA frame to the
server.

Fixes: QTBUG-123345
Pick-to: dev 6.7
Change-Id: I67a5fbacda435e121111e976aed1a140ef12bc7c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>