Add the 'writesDone' methods to client-side streaming operations
The 'writesDone' method ends the client-side streaming and notifies
server that the client finished the sending of data. Server still may
send the data to the client before terminating the communication. After
calling the method it's expected that the respective streaming
operation will emit either the finished signal. The core difference
between the 'cancel' and 'writesDone' methods is that stream processing
is not finished after the 'writesDone' method is called.

The QGrpcHttp2Client implements this logic by sending stream end
flag to a client-side HTTP/2 stream.

Task-number: QTBUG-123644
Task-number: QTBUG-123625
Change-Id: Ib648d02d167e8e630a7ef37816f62ce95c967273
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>